+++ test/Sentry.Tests/HubTests.cs
@@ -7,7 +7,6 @@ using NSubstitute;
 using Sentry;
 using Sentry.Extensibility;
 using Sentry.Internal;
-using Sentry.Protocol;
 using Sentry.Protocol.Envelopes;
 using Xunit;