+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add AddSentryTag and AddSentryContext Extensions for exception class (#834) @lucas-zimerman
+
 ## 3.0.7
 
 ### Changes
-- Add AddSentryTag and AddSentryContext Extensions for exception class (#834) @lucas-zimerman
 - Don't write timezone_display_name if it's the same as the ID (#837) @Tyrrrz
 - Serialize arbitrary objects in contexts (#838) @Tyrrrz