+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add AddSentryTag and AddSentryContext Extensions for exception class (#834) @lucas-zimerman
+- Associate span exceptions with event exceptions (#848) @Tyrrrz
 - MaxCacheItems option to control files on disk (#846) @Tyrrrz
 - Move SentryHttpMessageHandlerBuilderFilter to Sentry.Extensions.Logging (#845) @Tyrrrz